backend/requirements.txt
Flask==3.0.0
pandas==2.3.2
openpyxl==3.1.5
xlrd==2.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
geopy==2.4.1
googlemaps==4.10.0
requests==2.31.0
anthropic>=0.40.0
openai>=1.50.0
google-genai>=0.5.0